requirements.txt
# Web框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
starlette>=0.31.0

# LangChain相关
langchain>=0.0.300
langchain-core>=0.0.10
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.20

# 向量存储
chromadb>=0.4.18
faiss-cpu>=1.7.4

# Gradio UI
gradio>=4.0.0
aiofiles>=23.2.1
pillow>=10.0.0
websockets>=11.0.3

# 实用工具
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.3
httpx>=0.25.0
jinja2>=3.1.2
tiktoken>=0.5.1

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.25.0

# 安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 文档
typing-extensions>=4.7.0 